[FIX] runbot: append user bin dir to pre commands
When pip installs a package that contains a script, it goes in
`~/.local/bin` directory of the current user. This leads to a log warning
to warn the user that the script won't be accessible by other users.

e.g.:
`WARNING: The script pygmentize is installed in '/home/odoo/.local/bin'
which is not on PATH.`

This is particularly annoying for documentation builds because it may
confuse the users.

Also, it means that when a locally installed pip package overrides a
python package installed system wide (e.g.: by apt-get), it's still
globally script that runs when called.

With this commit, the user PATH is updated to include this local bin
directory when building Docker images.